requirements.txt
pyttsx3
SpeechRecognition
wikipedia
requests
geocoder
python-dotenv
playsound